apps/vibe-workflow/server/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
httpx
python-dotenv
ruff
